+++ src/index.js
@@ -5,6 +5,10 @@ import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './style/main.css'
 
+/** This is the entry point for the app. 
+ * The bulk of the app is in scene.js
+*/
+
 ReactDOM.render(
 	<ThreeScene />,
 	document.getElementById("root")
