+++ src/style/main.css
@@ -90,11 +90,6 @@ body
     margin:auto;
 }
 
-.camera-right
-{
-    margin-right:auto;
-}
-
 .scaleIcon
 {
     margin:1rem;
